Refactor : App.jsx에 있던 reducer와 state 파일로 분리
상세 :
- reducer를 store/globalReducer.js로 분리
- state를 store/globalState로 분리
