+++ client/src/App.jsx
@@ -3,30 +3,19 @@ import { Route, BrowserRouter as Router } from 'react-router-dom';
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core';
 import { MainPage, Game } from './presentation/pages';
 import { GlobalContext, DispatchContext } from './contexts';
+import { globalState, globalReducer } from './store';
 
 const App = () => {
-  const initialState = {
-    roomId: null,
-  };
   const theme = createMuiTheme({
     typography: { fontFamily: 'Aldrich' },
   });
 
-  const reducer = (state, action) => {
-    switch (action.type) {
-      case 'changeRoomId':
-        return { ...state, roomId: action.payload.roomId };
-      default:
-        throw new Error();
-    }
-  };
-
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const [state, dispatch] = useReducer(globalReducer, globalState);
 
   return (
     <MuiThemeProvider theme={theme}>
-      <GlobalContext.Provider value={{ state }}>
-        <DispatchContext.Provider value={{ dispatch }}>
+      <GlobalContext.Provider value={state}>
+        <DispatchContext.Provider value={dispatch}>
           <Router>
             <Route exact path="/" component={MainPage} />
             <Route path="/game" component={Game} />
