+++ client/src/store/globalState.js
@@ -1,0 +1,5 @@
+const initialState = {
+  roomId: null,
+};
+
+export default initialState;
