+++ client/src/store/index.js
@@ -1,0 +1,2 @@
+export { default as globalReducer } from './globalReducer';
+export { default as globalState } from './globalState';
